Add Existing Company User registration
Flow:
Add a New User to Existing Company
Sign out of Current Account
Log in to the New User Account
Complete Registration
Delete the New User

Note:
Using jenkins@risevision.com and his Company (account is used for Apps e2e tests) to add the new User.

The new User added is jenkins1@risevision.com which is used for e2e tests.

To prevent incorrect User and Company deletion, I've added checks for the Name of the Company to be deleted, and the Email of the User being deleted.